lib: Add dh_subst_multiarch script.
The new script is a modified dh_subst_env: one that only works on
DEB_(BUILD|HOST)_* variables, and uses dpkg-architecture to query
them, instead of turning to the environment.

dpkg-architecture will use an existing environment variable, if one
exists, so already exported variables will override whatever it'd
return otherwise. Due to this, we don't need to consult $ENV first.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>