dh-exec: Implement --with-scripts= option
The --with-scripts= option can be used to limit not only which helpers
get run, but which scripts they will try. Allows for a finer
granularity than the old --with option.

This is accomplished by assembling a DH_EXEC_SCRIPTS environment
variable that gets passed down to the helpers, and when those scan the
script directory, they will filter out anything that does not match
their part of the list. Unless the list is empty, of course.

Requested-by: Wookey <wookey@wookware.org>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>