Decouple input filtering from stripping
We need to do the filtering first, before other transformations are
applied - but stripping is done at the end, so that's no good. So this
introduces dh-exec-filter, which does the filtering.

As a side-effect, it also removes empty and commented lines before other
scripts are processed, thereby fixing Debian #757273.

However, as an unfortunate side effect, dh-exec now hangs if it does not
receive any input. This should be corrected one way or the other.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>