Do not hardcode the default script & helper paths
Instead of hardcoding the default DH_EXEC_SCRIPTDIR and DH_EXEC_LIBDIR
paths, regardless of any path-changing configure options, respect those,
and only fall back to the previously hardcoded values, if these
variables are not defined for some reason.

This makes it considerably easier to install and use dh-exec at a
non-default location.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>