dh-exec-filter-arch: Fix the multi-arch regression
In case the arch filter contains multiple architectures, dh-exec-filter
should still work. To this end, the way the filter is distinguished from
shell wildcards was changed yet again: now we require the filter to be
at the end of the line and be preceded by whitespace, or we require it
to be at the front, followed by space. Pretty much no restrictions on
content.

Also added a few new test cases, and as a side effect, this allows
filters such as `[amd64]` too, without the `linux-` prefix.

Closes: Debian#801716
Reported-by: Paul Gevers
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>