dh-exec-install-rename: Restore d/install support
With the DH_CONFIG_ACT_ON_PACKAGES change, dh-exec lost the ability to
work on debian/install and debian/manpages files, because those do not
have an explicit package name, it failed to recognise that it should
process them nevertheless.

To restore that functionality, we now check for bare "install" and
"manpages" names.

Closes: Debian#802034
Reported-by: Niko Tyni
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>